+++ backend/canisters/community/api/src/updates/c2c_join_channel.rs
@@ -2,7 +2,7 @@ use candid::{CandidType, Principal};
 use serde::{Deserialize, Serialize};
 use types::{
     ChannelId, CommunityCanisterChannelSummary, CommunityCanisterCommunitySummary, GateCheckFailedReason, TimestampMillis,
-    UniqueHumanProof, UserId, VerifiedCredentialGateArgs,
+    UniquePersonProof, UserId, VerifiedCredentialGateArgs,
 };
 
 #[derive(CandidType, Serialize, Deserialize, Debug)]
@@ -15,7 +15,7 @@ pub struct Args {
     pub is_bot: bool,
     pub diamond_membership_expires_at: Option<TimestampMillis>,
     pub verified_credential_args: Option<VerifiedCredentialGateArgs>,
-    pub unique_human_proof: Option<UniqueHumanProof>,
+    pub unique_person_proof: Option<UniquePersonProof>,
 }
 
 #[derive(CandidType, Serialize, Deserialize, Debug)]