+++ backend/canisters/community/api/src/updates/c2c_join_community.rs
@@ -1,7 +1,7 @@
 use candid::{CandidType, Principal};
 use serde::{Deserialize, Serialize};
 use types::{
-    CommunityCanisterCommunitySummary, GateCheckFailedReason, TimestampMillis, UniqueHumanProof, UserId,
+    CommunityCanisterCommunitySummary, GateCheckFailedReason, TimestampMillis, UniquePersonProof, UserId,
     VerifiedCredentialGateArgs,
 };
 
@@ -14,7 +14,7 @@ pub struct Args {
     pub is_bot: bool,
     pub diamond_membership_expires_at: Option<TimestampMillis>,
     pub verified_credential_args: Option<VerifiedCredentialGateArgs>,
-    pub unique_human_proof: Option<UniqueHumanProof>,
+    pub unique_person_proof: Option<UniquePersonProof>,
 }
 
 #[derive(CandidType, Serialize, Deserialize, Debug)]