+++ backend/canisters/community/impl/src/updates/create_channel.rs
@@ -48,7 +48,7 @@ fn c2c_create_proposals_channel(args: Args) -> Response {
                 is_bot: true,
                 diamond_membership_expires_at: None,
                 verified_credential_args: None,
-                unique_human_proof: None,
+                unique_person_proof: None,
             },
             state,
         )