+++ backend/canisters/group/api/src/updates/c2c_join_group.rs
@@ -1,7 +1,8 @@
 use candid::{CandidType, Principal};
 use serde::{Deserialize, Serialize};
 use types::{
-    GateCheckFailedReason, GroupCanisterGroupChatSummary, TimestampMillis, UniqueHumanProof, UserId, VerifiedCredentialGateArgs,
+    GateCheckFailedReason, GroupCanisterGroupChatSummary, TimestampMillis, UniquePersonProof, UserId,
+    VerifiedCredentialGateArgs,
 };
 
 #[derive(CandidType, Serialize, Deserialize, Debug)]
@@ -14,7 +15,7 @@ pub struct Args {
     pub is_bot: bool,
     pub diamond_membership_expires_at: Option<TimestampMillis>,
     pub verified_credential_args: Option<VerifiedCredentialGateArgs>,
-    pub unique_human_proof: Option<UniqueHumanProof>,
+    pub unique_person_proof: Option<UniquePersonProof>,
 }
 
 #[derive(CandidType, Serialize, Deserialize, Debug)]