+++ backend/canisters/local_user_index/impl/src/updates/join_channel.rs
@@ -18,7 +18,7 @@ async fn join_channel(args: Args) -> Response {
         is_bot: user_details.is_bot,
         diamond_membership_expires_at: user_details.diamond_membership_expires_at,
         verified_credential_args: args.verified_credential_args.clone(),
-        unique_human_proof: user_details.unique_human_proof.clone(),
+        unique_person_proof: user_details.unique_person_proof.clone(),
     };
     match community_canister_c2c_client::c2c_join_channel(args.community_id.into(), &c2c_args).await {
         Ok(response) => match response {