+++ backend/canisters/user_index/impl/src/model/user.rs
@@ -4,7 +4,7 @@ use candid::{CandidType, Principal};
 use serde::{Deserialize, Serialize};
 use types::{
     is_default, is_empty_slice, CyclesTopUp, CyclesTopUpInternal, PhoneNumber, RegistrationFee, SuspensionAction,
-    SuspensionDuration, TimestampMillis, UniqueHumanProof, UserId, UserSummary, UserSummaryStable, UserSummaryV2,
+    SuspensionDuration, TimestampMillis, UniquePersonProof, UserId, UserSummary, UserSummaryStable, UserSummaryV2,
     UserSummaryVolatile,
 };
 
@@ -61,7 +61,7 @@ pub struct User {
     #[serde(rename = "lc", default)]
     pub latest_chit_event: TimestampMillis,
     #[serde(rename = "uh", default, skip_serializing_if = "Option::is_none")]
-    pub unique_human_proof: Option<UniqueHumanProof>,
+    pub unique_person_proof: Option<UniquePersonProof>,
 }
 
 impl User {
@@ -116,7 +116,7 @@ impl User {
             streak: 0,
             streak_ends: 0,
             latest_chit_event: 0,
-            unique_human_proof: None,
+            unique_person_proof: None,
         }
     }
 
@@ -228,7 +228,7 @@ impl Default for User {
             streak_ends: 0,
             chit_updated: 0,
             latest_chit_event: 0,
-            unique_human_proof: None,
+            unique_person_proof: None,
         }
     }
 }