+++ backend/libraries/types/src/proof_of_uniqueness.rs
@@ -3,12 +3,12 @@ use candid::CandidType;
 use serde::{Deserialize, Serialize};
 
 #[derive(CandidType, Serialize, Deserialize, Clone, Debug)]
-pub struct UniqueHumanProof {
+pub struct UniquePersonProof {
     pub timestamp: TimestampMillis,
-    pub provider: UniqueHumanProofProvider,
+    pub provider: UniquePersonProofProvider,
 }
 
 #[derive(CandidType, Serialize, Deserialize, Clone, Debug)]
-pub enum UniqueHumanProofProvider {
+pub enum UniquePersonProofProvider {
     DecideAI,
 }